requirements.txt
pandas==2.2.1
python-dotenv==1.0.0
openpyxl==3.1.2